+++ packages/vhd-lib/peekFooterFromVhdStream.js
@@ -3,11 +3,10 @@
 const { readChunk } = require('@vates/read-chunk')
 
 const { FOOTER_SIZE } = require('./_constants')
-const { fuFooter } = require('./_structs')
+const { unpackFooter } = require('./Vhd/_utils.js')
 
 module.exports = async function peekFooterFromStream(stream) {
-  const footerBuffer = await readChunk(stream, FOOTER_SIZE)
-  const footer = fuFooter.unpack(footerBuffer)
-  stream.unshift(footerBuffer)
-  return footer
+  const buffer = await readChunk(stream, FOOTER_SIZE)
+  stream.unshift(buffer)
+  return unpackFooter(buffer)
 }