+++ packages/xo-web/src/xo-app/backup/new/index.js
@@ -299,7 +299,9 @@ const New = decorate([
               newSchedule.cron !== oldSchedule.cron ||
               newSchedule.name !== oldSchedule.name ||
               newSchedule.timezone !== oldSchedule.timezone ||
-              newSchedule.enabled !== oldSchedule.enabled
+              newSchedule.enabled !== oldSchedule.enabled ||
+              newSchedule.healthCheckSr !== oldSchedule.healthCheckSr ||
+              newSchedule.healthCheckVmsWithTags !== oldSchedule.healthCheckVmsWithTags
             ) {
               return editSchedule({
                 id,
@@ -307,6 +309,8 @@ const New = decorate([
                 name: newSchedule.name,
                 timezone: newSchedule.timezone,
                 enabled: newSchedule.enabled,
+                healthCheckSr: newSchedule.healthCheckSr,
+                healthCheckVmsWithTags: newSchedule.healthCheckVmsWithTags,
               })
             }
           })
@@ -322,6 +326,8 @@ const New = decorate([
                 name: schedule.name,
                 timezone: schedule.timezone,
                 enabled: schedule.enabled,
+                healthCheckSr: schedule.healthCheckSr,
+                healthCheckVmsWithTags: schedule.healthCheckVmsWithTags,
               })
 
               settings = settings.withMutations(settings => {
@@ -488,7 +494,19 @@ const New = decorate([
       saveSchedule:
         (
           _,
-          { copyRetention, cron, enabled = true, exportRetention, fullInterval, id, name, snapshotRetention, timezone }
+          {
+            copyRetention,
+            cron,
+            enabled = true,
+            exportRetention,
+            fullInterval,
+            healthCheckSr,
+            healthCheckVmsWithTags,
+            id,
+            name,
+            snapshotRetention,
+            timezone,
+          }
         ) =>
         ({ propSettings, schedules, settings = propSettings }) => ({
           schedules: {
@@ -497,6 +515,8 @@ const New = decorate([
               ...schedules[id],
               cron,
               enabled,
+              healthCheckSr,
+              healthCheckVmsWithTags,
               id,
               name,
               timezone,